backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["caitlynfowkes <caitlynfowkes@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-allauth = "^64.1.0"
dj-rest-auth = "^6.0.0"
django-cors-headers = "^4.4.0"
requests = "^2.32.3"
psycopg2 = "^2.9.10"
django-extensions = "^3.2.3"
django-environ = "^0.11.2"
pillow = "^11.1.0"
factory-boy = "^3.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
